backend/requirements.txt
flask
flask-cors
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-core
faiss-cpu
sentence-transformers
groq
pandas
gunicorn
python-dotenv